Updated the launchpad API used to allow default config builds
Now if no architecture is specified it will trigger a rebuild using the
default launchpad configuration which selects the serie according to the
snapcraft.yaml content and trigger rebuilds on the active architectures.

It's still possible to trigger a specific build by specifying an arch,
in which case the series used isn't automatic anymore.